Authorization assignment completed

+++ src/controllers/postController.js
@@ -4,7 +4,6 @@ const Authorizer = require('../policies/post');
 module.exports = {
     new(req, res, next) {
         const authorized = new Authorizer(req.user).new();
-        console.log('req.user within new postcontroller: ' + req.user.role)
         if(authorized) {
             res.render('posts/new', {topicId: req.params.topicId});
         } else {
@@ -14,7 +13,6 @@ module.exports = {
     },
     create(req, res, next) {
         const authorized = new Authorizer(req.user).create();
-        console.log('req.user within create postcontroller: ' + req.user.role);
         if(authorized) {
             let newPost = {
                 title: req.body.title,
@@ -45,30 +43,53 @@ module.exports = {
         });
     },
     destroy(req, res, next) {
-        postQueries.deletePost(req.params.id, (err, deletedRecordsCount) => {
-            if(err) {
-                res.redirect(500, `/topics/${req.params.topicId}/posts/${req.params.id}`);
-            } else {
-                res.redirect(303, `/topics/${req.params.topicId}`);
-            }
-        });
+        const authorized = new Authorizer(req.user).destroy();
+
+        if (authorized) {
+            postQueries.deletePost(req.params.id, (err, deletedRecordsCount) => {
+                if(err) {
+                    res.redirect(500, `/topics/${req.params.topicId}/posts/${req.params.id}`);
+                } else {
+                    res.redirect(303, `/topics/${req.params.topicId}`);
+                }
+            });
+        } else {
+            req.flash('notice', 'You are not authorized to do that.');
+            res.redirect('/posts');
+        }
+        
     },
     edit(req, res, next) {
-        postQueries.getPost(req.params.id, (err, post) => {
-            if(err || post == null) {
-                res.redirect(404, '/');
-            } else {
-                res.render('posts/edit', {post});
-            }
-        });
+        const authorized = new Authorizer(req.user).edit();
+        if (authorized) {
+            postQueries.getPost(req.params.id, (err, post) => {
+                if(err || post == null) {
+                    res.redirect(404, '/');
+                } else {
+                    res.render('posts/edit', {post});
+                }
+            });
+        } else {
+            req.flash('notice', 'You are not authorized to do that.');
+            res.redirect('/posts');
+        }
+        
     },
     update(req, res, next) {
-        postQueries.updatePost(req, req.body, (err, post) => {
-            if(err || post == null) {
-                res.redirect(404, `/topics/${req.params.topicId}/posts/${req.params.id}/edit`);
-            } else {
-                res.redirect(`/topics/${req.params.topicId}/posts/${req.params.id}`);
-            }
-        });
+        const authorized = new Authorizer(req.user).update();
+
+        if(authorized) {
+            postQueries.updatePost(req, req.body, (err, post) => {
+                if(err || post == null) {
+                    res.redirect(404, `/topics/${req.params.topicId}/posts/${req.params.id}/edit`);
+                } else {
+                    res.redirect(`/topics/${req.params.topicId}/posts/${req.params.id}`);
+                }
+            });
+        } else {
+            req.flash('notice', 'You are not authorized to do that.');
+            res.redirect('/posts');
+        }
+        
     }
 }
